chore: define non-default http-client features
